requirements.txt
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.38.6
sqlalchemy==2.0.36
python-multipart==0.0.12
bcrypt==4.2.1
qrcode==8.0
pillow==10.4.0
requests==2.32.3
cryptography==43.0.3
apscheduler==3.10.4
